prevent multiple web server installations
