+++ app/SSH/Services/Webserver/Nginx.php
@@ -6,7 +6,6 @@ use App\Exceptions\SSHError;
 use App\Exceptions\SSLCreationException;
 use App\Models\Site;
 use App\Models\Ssl;
-use Closure;
 use Throwable;
 
 class Nginx extends AbstractWebserver
@@ -32,21 +31,6 @@ class Nginx extends AbstractWebserver
         $this->service->server->systemd()->restart('nginx');
 
         $this->service->server->os()->cleanup();
-    }
-
-    public function deletionRules(): array
-    {
-        return [
-            'service' => [
-                function (string $attribute, mixed $value, Closure $fail): void {
-                    $hasSite = $this->service->server->sites()
-                        ->exists();
-                    if ($hasSite) {
-                        $fail('Cannot uninstall webserver while you have websites using it.');
-                    }
-                },
-            ],
-        ];
     }
 
     /**
